@@ -1,3 +1,3 @@
 module ethohampton.com/BallotValidator
 
-go 1.17
+go 1.23
