@@ -5,6 +5,7 @@ import (
 	"hash/fnv"
 	"log"
 	"math/rand"
+	"slices"
 
 	"ethohampton.com/BallotValidator/internal/util"
 )
@@ -30,7 +31,7 @@ func StepFour(previousVotes []string, votes []util.Vote, seed string, numberToPi
 		//otherwise it is in the previous votes, so use that
 		if randomVal < initialSizeCurrentVotes {
 			winner := votes[randomVal] // pick winner
-			if !util.Contains(&winners, winner.ONID) {
+			if !slices.Contains(winners, winner.ONID) {
 				winners = append(winners, winner.ONID)
 				logMessages = append(logMessages, "Winner: "+winner.ONID+" with response ID "+winner.ID+" chosen with random value "+fmt.Sprint(randomVal))
 			} else {
@@ -38,7 +39,7 @@ func StepFour(previousVotes []string, votes []util.Vote, seed string, numberToPi
 			}
 		} else if randomVal < initialSizeCurrentVotes+initialSizePreviousVotes {
 			winner := previousVotes[randomVal-initialSizeCurrentVotes] // pick winner
-			if !util.Contains(&winners, winner) {
+			if !slices.Contains(winners, winner) {
 				winners = append(winners, winner)
 				logMessages = append(logMessages, "Winner: "+winner+" chosen from past responses with random value "+fmt.Sprint(randomVal))
 			} else {
