@@ -2,24 +2,6 @@ package util
 
 import "testing"
 
-// never hurt to do some sanity checks on the most basic stuff
-func TestContains(t *testing.T) {
-	var testCases = []struct {
-		slice  []string
-		value  string
-		expect bool
-	}{
-		{[]string{"a", "b", "c"}, "a", true},
-		{[]string{"a", "b", "c"}, "d", false},
-	}
-
-	for _, testCase := range testCases {
-		if Contains(&testCase.slice, testCase.value) != testCase.expect {
-			t.Errorf("Expected %v to be %v", testCase.slice, testCase.expect)
-		}
-	}
-}
-
 func TestRemoveDuplicateStr(t *testing.T) {
 	var testCases = []struct {
 		slice  []string
@@ -43,7 +25,7 @@ func TestRemoveDuplicateStr(t *testing.T) {
 	}
 }
 
-func TestCleanVote(t *testing.T) {
+func TestNormalizeVote(t *testing.T) {
 	var testCases = []struct {
 		vote   string
 		expect string
